changed Hash type to uint32

+++ src/shared/GeneralTypes.h
@@ -7,8 +7,8 @@
 #define SERVER_DEFAULT_PORT 59095
 #define CLIENT_DEFAULT_PORT 59096
 #define DEFAULTSEGMENTSIZE 8000
-typedef uint64_t Id;
-typedef uint64_t Hash;
+typedef u_int32_t Id;
+typedef u_int32_t Hash;
 typedef std::string IpV4Address;
 typedef std::string IpV6Address;
 typedef int Port;
